refactor(anyharness): split api contract boundary modules
