+++ anyharness/crates/anyharness-lib/src/api/http/mod.rs
@@ -8,9 +8,13 @@ pub mod cowork;
 pub mod error;
 pub mod files;
 pub mod git;
+mod git_contract;
+mod git_task;
 pub mod health;
 pub mod hosting;
 pub mod mobility;
+mod mobility_archive_contract;
+mod mobility_contract;
 pub mod plans;
 pub mod processes;
 pub mod product_mcp;
@@ -19,8 +23,23 @@ pub mod repo_roots;
 pub mod reviews;
 pub mod runtime_config;
 pub mod sessions;
+pub mod sessions_config;
+mod sessions_contract;
+mod sessions_errors;
+pub mod sessions_events;
+pub mod sessions_fork;
+pub mod sessions_interactions;
+mod sessions_leases;
+pub mod sessions_lifecycle;
+pub mod sessions_pending;
+pub mod sessions_prompt;
+pub mod sessions_resume;
 pub mod subagents;
 pub mod terminals;
 pub mod workspaces;
 pub(crate) mod workspaces_contract;
+pub mod workspaces_lifecycle;
+pub mod workspaces_purge;
+pub mod workspaces_setup;
+pub mod workspaces_worktrees;
 pub mod worktrees;
