+++ anyharness/crates/anyharness-lib/src/api/openapi.rs
@@ -179,52 +179,52 @@ use anyharness_contract::v1::{
         super::http::cowork::get_cowork_artifact,
         super::http::workspaces::resolve_workspace,
         super::http::workspaces::create_workspace,
-        super::http::workspaces::create_worktree,
+        super::http::workspaces_worktrees::create_worktree,
         super::http::workspaces::list_workspaces,
         super::http::workspaces::get_workspace,
-        super::http::workspaces::retire_workspace_preflight,
-        super::http::workspaces::retire_workspace,
-        super::http::workspaces::retry_retire_cleanup,
-        super::http::workspaces::purge_workspace_preflight,
-        super::http::workspaces::purge_workspace,
-        super::http::workspaces::retry_purge_workspace,
+        super::http::workspaces_lifecycle::retire_workspace_preflight,
+        super::http::workspaces_lifecycle::retire_workspace,
+        super::http::workspaces_lifecycle::retry_retire_cleanup,
+        super::http::workspaces_purge::purge_workspace_preflight,
+        super::http::workspaces_purge::purge_workspace,
+        super::http::workspaces_purge::retry_purge_workspace,
         super::http::worktrees::get_worktree_inventory,
         super::http::worktrees::prune_orphan_worktree,
         super::http::worktrees::get_worktree_retention_policy,
         super::http::worktrees::update_worktree_retention_policy,
         super::http::worktrees::run_worktree_retention,
         super::http::workspaces::update_workspace_display_name,
-        super::http::workspaces::detect_project_setup,
-        super::http::workspaces::get_setup_status,
-        super::http::workspaces::rerun_setup,
-        super::http::workspaces::start_setup,
+        super::http::workspaces_setup::detect_project_setup,
+        super::http::workspaces_setup::get_setup_status,
+        super::http::workspaces_setup::rerun_setup,
+        super::http::workspaces_setup::start_setup,
         super::http::mobility::preflight_workspace_mobility,
         super::http::mobility::update_workspace_mobility_runtime_state,
         super::http::mobility::export_workspace_mobility_archive,
         super::http::mobility::install_workspace_mobility_archive,
         super::http::mobility::destroy_workspace_mobility_source,
         super::http::sessions::create_session,
-        super::http::sessions::update_session_title,
-        super::http::sessions::get_live_session_config,
-        super::http::sessions::set_session_config_option,
-        super::http::sessions::prompt_session,
-        super::http::sessions::fork_session,
-        super::http::sessions::edit_pending_prompt,
-        super::http::sessions::delete_pending_prompt,
-        super::http::sessions::get_prompt_attachment,
-        super::http::sessions::resume_session,
-        super::http::sessions::cancel_session,
-        super::http::sessions::close_session,
-        super::http::sessions::dismiss_session,
-        super::http::sessions::restore_dismissed_session,
+        super::http::sessions_config::update_session_title,
+        super::http::sessions_config::get_live_session_config,
+        super::http::sessions_config::set_session_config_option,
+        super::http::sessions_prompt::prompt_session,
+        super::http::sessions_fork::fork_session,
+        super::http::sessions_pending::edit_pending_prompt,
+        super::http::sessions_pending::delete_pending_prompt,
+        super::http::sessions_pending::get_prompt_attachment,
+        super::http::sessions_resume::resume_session,
+        super::http::sessions_lifecycle::cancel_session,
+        super::http::sessions_lifecycle::close_session,
+        super::http::sessions_lifecycle::dismiss_session,
+        super::http::sessions_lifecycle::restore_dismissed_session,
         super::http::sessions::list_sessions,
         super::http::sessions::get_session,
         super::http::subagents::get_session_subagents,
         super::http::subagents::schedule_subagent_wake,
-        super::http::sessions::list_session_events,
-        super::http::sessions::list_session_raw_notifications,
-        super::http::sessions::resolve_interaction,
-        super::http::sessions::reveal_mcp_elicitation_url,
+        super::http::sessions_events::list_session_events,
+        super::http::sessions_events::list_session_raw_notifications,
+        super::http::sessions_interactions::resolve_interaction,
+        super::http::sessions_interactions::reveal_mcp_elicitation_url,
         super::http::replay::list_replay_recordings,
         super::http::replay::export_replay_recording,
         super::http::replay::create_replay_session,
